Rename gree coordinator module
